device/requirements.txt
awscrt==0.23.5
awsiotsdk==1.22.1
protobuf==5.29.3
